requirements.txt
NoiseProtocol>=0.3.1
argparse>=1.4
prompt-toolkit>=3.0.5